Draw chevrons based on the current time view.
Sometimes we have an individual whose temporal extent is bounded, but
those bounds are outside our current view. This happens when we are
viewing a sub-activity and an individual starts or ends in a different
activity outside our view.

The chevrons simply mean 'this individual exists before/after the piece
of time we are showing here'. So check against those bounds.